+++ CHANGELOG.md
@@ -6,7 +6,9 @@
 - Add `--color` option to `autumn dump-tree-sitter` command for colored AST output
 
 ### Changed
-- Update parsers: angular, latex, llvm, vim
+- Sync vendored parsers with nvim-treesitter repo
+- Add language markdown-inline
+- Update parsers: angular, latex, llvm, markdown, perl, vim
 - Update queries: c, csharp, ecma, fsharp, javascript, php, powershell, swift, tsx
 - Improve `autumn dump-tree-sitter` output to display field names and match Neovim's `:InspectTree` format while preserving raw text tokens